always send both axis of a stick when sending a gamepadChange event
This prevents axis active state from being false when one axis is at the border and other axis is at the center, or when the stick is pushed all the way to the end of single axis.
The event is still only dispatched when the timestamp changes, so this change won't make the event firing the all time.