fix jittering of LS on joystick skin when dpad on gamepad is in use
Skins with 'joystick' properties will prioritize dpad over sticks now.
This is not a problem for joysticks which will only have one of either three active: LS, RS, and dpad.
This is also preferable to gamepads which may have sensitive sticks, because sticks are really not used when the user wants to use joystick skins on their gamepad.

Also joysticks are just gamepads to Gamepad API. Unless specified, there's no way to tell if they're generic ones or joysticks.