make button value off check flexible
It will always check the value's two members, then the value itself.
If the value is a number (most of the cases), it will first always be 'off' then second phrase will properly check if it's off, and correct the final result.
If the value is an array of two numbers (dpad value), it will be properly checked by the first phrase, skipping the second phrase.

This change should make the check always pass the proper result regardless of the button value's type, making it possible to make a skin that uses dpad value and left analog stick at the same time.